% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/planets.R
\docType{data}
\name{planets}
\alias{planets}
\alias{planets-package}
\title{"Planets"}
\format{A data frame with 3915 observations of 12 variables}
\usage{
planets
}
\description{
A data set containing the most basic information available of all known planets.

This package contains basic information from all knwon Planets
}
\examples{
data(planets)

plot(
data = planets,
`Mass (MJ)` ~ `Temperature (K)`
)
}
\keyword{datasets}
